refactor: organize utility modules
