+++ src/ChangeLog.tsx
@@ -2,8 +2,8 @@ import { Fragment, useEffect, useMemo, useRef, useState } from "react";
 import ItemImg from "./ItemImg";
 import SearchBar from "./shared/SearchBar";
 import { useContextMenu, CtxMenu } from "./shared/CtxMenu";
-import { openWiki, copyWikiLink } from "./wiki";
-import { formatUnixTime } from "./formatters";
+import { openWiki, copyWikiLink } from "./lib/wiki";
+import { formatUnixTime } from "./lib/formatters";
 import type { ClockFormat } from "./types/settings";
 import type { ChangeLogEntry } from "./types/inventory";
 import "./ChangeLog.css";
