+++ src/lib/formatters.ts
@@ -1,4 +1,4 @@
-import type { ClockFormat } from "./types/settings";
+import type { ClockFormat } from "../types/settings";
 
 export function formatBytes(bytes: number): string {
   if (bytes >= 1024 * 1024) return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
